+++ src/index.ts
@@ -1,3 +1,3 @@
 export * from './calculation';
 
-export const version = '1.1.0-dev';
+export const version = '1.1.0-dev.001';